feat: User feedback for subscription cancellation and reactivation

+++ app/src/components/features/subscription/CancelSubscription.component.tsx
@@ -8,41 +8,62 @@ import {
 import { appConfig } from '@/config'
 import supabase from '@/services/api/supabase'
 import { useSubscription } from '@/services/context/SubscriptionContext'
+import { useUser } from '@/services/context/UserContext'
+import fetchErrorToast from '@/hooks/fetchErrorToast'
+import { toast } from 'sonner'
+import { useState } from 'react'
 
 interface CancelSubscriptionProps {
   onCloseModal?: () => void
 }
 
 function CancelSubscription({ onCloseModal }: CancelSubscriptionProps) {
-  const { subscription } = useSubscription()
-  async function handleDelete() {
-    const {
-      data: { session },
-    } = await supabase.auth.getSession()
-    const token = session?.access_token
+  const [status, setStatus] = useState<'IDLE' | 'LOADING' | 'ERROR'>('IDLE')
+  const { user } = useUser()
+  const { subscription, getSubscription } = useSubscription()
 
-    const res = await fetch(
-      `${appConfig.apiUrl}/subscriptions/${subscription?.stripe_subscription_id}`,
-      {
-        method: 'DELETE',
-        headers: {
-          Authorization: `Bearer ${token}`,
-          'Content-Type': 'application/json',
+  async function handleDelete() {
+    if (!user) return
+    setStatus('LOADING')
+    try {
+      const {
+        data: { session },
+      } = await supabase.auth.getSession()
+      const token = session?.access_token
+
+      const res = await fetch(
+        `${appConfig.apiUrl}/subscriptions/${subscription?.stripe_subscription_id}`,
+        {
+          method: 'DELETE',
+          headers: {
+            Authorization: `Bearer ${token}`,
+            'Content-Type': 'application/json',
+          },
         },
-      },
-    )
-    const data = await res.json()
-    console.log(data)
+      )
+      const data = await res.json()
+      if (data.status !== 'success') throw new Error()
+
+      await getSubscription(user?.id)
+      toast.info('Abo erfolgreich gekündigt.')
+      setStatus('IDLE')
+      onCloseModal?.()
+    } catch (e) {
+      setStatus('ERROR')
+    }
   }
+
   return (
     <div>
       <DialogHeader>
-        <DialogTitle>Möchtest du dein Abo wirklich kündigen?</DialogTitle>
+        <DialogTitle>Möchtest du dein Abo wirklich beenden?</DialogTitle>
       </DialogHeader>
 
       <DialogDescription>
-        Dein Abo ist nach der Kündigung noch bis zum Ende der Laufzeit aktiv. Du
-        kannst alle erfassten Daten auch nach der Kündigung jederzeit einsehen.
+        Dein Abo ist nach der Kündigung noch bis zum Ende der Laufzeit aktiv.
+        Danach kannst du keine neuen Daten erfassen. Alle bereits erfassten
+        Daten sind aber auch nach Ende der Laufzeit jederzeit für dich
+        verfügbar.
       </DialogDescription>
       <div className='flex justify-end gap-4 mt-4'>
         <Button size='sm' variant='outline' onClick={onCloseModal}>
@@ -50,21 +71,21 @@ function CancelSubscription({ onCloseModal }: CancelSubscriptionProps) {
         </Button>
         <div className='flex items-center gap-2'>
           <Button
-            // disabled={}
+            disabled={status === 'LOADING'}
             size='sm'
             variant='destructive'
             onClick={handleDelete}
           >
-            Abo kündigen
+            Abo beenden
           </Button>
-          {/* {isDeleting && <MiniLoader />} */}
+          {status === 'LOADING' && <MiniLoader />}
         </div>
       </div>
-      {/* {isError && ( */}
-      {/*   <p className='mt-4 text-center text-sm text-warning'> */}
-      {/*     Es ist etwas schiefgelaufen. Versuch's nochmal. */}
-      {/*   </p> */}
-      {/* )} */}
+      {status === 'ERROR' && (
+        <p className='mt-4 text-center text-sm text-warning'>
+          Es ist etwas schiefgelaufen. Versuch's nochmal.
+        </p>
+      )}
     </div>
   )
 }
